Moved notifications back to top level because it might break some installations
