+++ script/join.py
